// DXC_Frontend/src/api/models/ocopEnterpriseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface OcopEnterpriseDto {
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  phoneNumber?: string | null
  /** @nullable */
  representative?: string | null
  /** @nullable */
  taxCode?: string | null
  /** @nullable */
  establishedYear?: number | null
  /** @nullable */
  address?: string | null
  /** @nullable */
  ocopCertificateNumber?: string | null
  /** @nullable */
  latitude?: number | null
  /** @nullable */
  longitude?: number | null
  isActive?: boolean
  createdAt?: string
  /** @nullable */
  updatedAt?: string | null
}